requirements.txt
# Core dependencies for Agentic RAG Masterclass
# Python 3.11+

# Interactive Development
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0

# LLM & Embeddings
openai>=1.12.0

# Agent Framework
langgraph>=0.0.28
langchain>=0.1.0
langchain-core>=0.1.0

# Vector Store
faiss-cpu>=1.7.4
numpy>=1.24.0

# Data Validation
pydantic>=2.5.0

# Evaluation (optional for advanced demos)
ragas>=0.1.0
datasets>=2.16.0

# API & Deployment
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
structlog>=24.1.0

# Optional: Observability
# opentelemetry-api>=1.22.0
# opentelemetry-sdk>=1.22.0
# prometheus-client>=0.19.0
